Combined new fhirpath methods with existing implementation
