+++ shared/src/main/java/gov/hhs/cdc/trustedintermediary/external/hapi/HapiFhirImplementation.java
@@ -1,17 +1,25 @@
 package gov.hhs.cdc.trustedintermediary.external.hapi;
 
 import ca.uhn.fhir.context.FhirContext;
+import ca.uhn.fhir.fhirpath.IFhirPath;
 import ca.uhn.fhir.parser.IParser;
 import gov.hhs.cdc.trustedintermediary.wrappers.FhirParseException;
 import gov.hhs.cdc.trustedintermediary.wrappers.HapiFhir;
+import gov.hhs.cdc.trustedintermediary.wrappers.formatter.Formatter;
+import java.util.List;
+import javax.inject.Inject;
 import org.hl7.fhir.instance.model.api.IBaseResource;
+import org.hl7.fhir.r4.model.Base;
+import org.hl7.fhir.r4.model.BooleanType;
 
 /** Concrete implementation that calls the Hapi FHIR library. */
 public class HapiFhirImplementation implements HapiFhir {
 
     private static final HapiFhirImplementation INSTANCE = new HapiFhirImplementation();
-
     private static final FhirContext CONTEXT = FhirContext.forR4();
+    private static final IFhirPath pathEngine = CONTEXT.newFhirPath();
+
+    @Inject Formatter formatter;
 
     private HapiFhirImplementation() {}
 
@@ -37,5 +45,15 @@ public class HapiFhirImplementation implements HapiFhir {
     public String encodeResourceToJson(Object resource) {
         IParser encodeResourceParser = CONTEXT.newJsonParser();
         return encodeResourceParser.encodeResourceToString((IBaseResource) resource);
+    }
+
+    public List<Base> evaluate(IBaseResource root, String expression) {
+        return pathEngine.evaluate(root, expression, Base.class);
+    }
+
+    @Override
+    public Boolean evaluateCondition(IBaseResource root, String expression) {
+        var result = pathEngine.evaluateFirst(root, expression, BooleanType.class);
+        return result.map(BooleanType::booleanValue).orElse(false);
     }
 }
