+++ shared/src/main/java/gov/hhs/cdc/trustedintermediary/wrappers/HapiFhir.java
@@ -1,6 +1,8 @@
 package gov.hhs.cdc.trustedintermediary.wrappers;
 
+import java.util.List;
 import org.hl7.fhir.instance.model.api.IBaseResource;
+import org.hl7.fhir.r4.model.Base;
 
 /**
  * An interface that wraps around the Hapi FHIR library. It is Hapi-specific because making it
@@ -13,4 +15,8 @@ public interface HapiFhir {
             throws FhirParseException;
 
     String encodeResourceToJson(Object resource);
+
+    List<Base> evaluate(IBaseResource root, String expression);
+
+    Boolean evaluateCondition(IBaseResource root, String expression);
 }
